Restructure roles and groups components
ref #3967
