+++ portal/src/components/roles-and-groups/dialog/DeleteRoleGroupDialog.tsx
@@ -1,9 +1,9 @@
 import React, { useCallback, useContext } from "react";
 import { Context } from "@oursky/react-messageformat";
-import { useRemoveRoleFromGroupsMutation } from "../../graphql/adminapi/mutations/removeRoleFromGroups";
-import { useRoleQuery } from "../../graphql/adminapi/query/roleQuery";
-import { useSnapshotData } from "../../hook/useSnapshotData";
-import RolesAndGroupsBaseDeleteDialog from "./RolesAndGroupsBaseDeleteDialog";
+import { useRemoveRoleFromGroupsMutation } from "../../../graphql/adminapi/mutations/removeRoleFromGroups";
+import { useRoleQuery } from "../../../graphql/adminapi/query/roleQuery";
+import { useSnapshotData } from "../../../hook/useSnapshotData";
+import RolesAndGroupsBaseDeleteDialog from "./components/RolesAndGroupsBaseDeleteDialog";
 
 export interface DeleteRoleGroupDialogData {
   roleID: string;
