+++ portal/src/components/roles-and-groups/form/RoleAndGroupsFormContainer.tsx
@@ -4,10 +4,10 @@ import {
   FormContainerBaseProps,
   FormModel,
   useFormContainerBaseContext,
-} from "../../FormContainerBase";
-import { useFormTopErrors } from "../../form";
-import { useErrorMessageBarContext } from "../../ErrorMessageBar";
-import { useLoading } from "../../hook/loading";
+} from "../../../FormContainerBase";
+import { useFormTopErrors } from "../../../form";
+import { useErrorMessageBarContext } from "../../../ErrorMessageBar";
+import { useLoading } from "../../../hook/loading";
 
 export interface RoleAndGroupsFormContainerProps<Form = FormModel>
   extends FormContainerBaseProps<Form> {}
